+++ tests/test_apps/cloudflare-workers/vitest.config.ts
@@ -1,0 +1,17 @@
+import { defineWorkersConfig } from "@cloudflare/vitest-pool-workers/config";
+
+export default defineWorkersConfig({
+	test: {
+		globals: true,
+		poolOptions: {
+			workers: {
+				main: "./tests/worker.ts",
+				wrangler: {
+					configPath: "./wrangler.toml",
+				},
+			},
+		},
+		testTimeout: 60000,
+		include: ["tests/**/*.spec.ts"],
+	},
+});